Remove broken JSON login test case
The test did not work correctly because the admin login view
does not accept JSON input, but instead produces an error on JSON POST.

It appeared to be functional and working because the
monkey-patched decorator returned a HTTP 403 when a user was accessing the view.

This did not actually result from a failed login,
but from a premature view permission checking that was not sane.

The HTTP 403 was produced by the login function decorator
before the user was ever accessing or using the actual view.

After modifying the view checking logic to run on a login attempt
instead of before attempt (ie when accessing the view)
the admin view error that resulted from missing form POST data
was discovered and since no easy fix was available without
writing a custom view the broken test case was removed altogether.

The correct fix for supporting JSON payloads and mimetypes
can be implemented by writing a login view that supports
JSON login via eg HTTP AJAX POST and produces the same errors.

Signed-off-by: Aleksi Häkli <aleksi.hakli@iki.fi>